auth/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
pydantic==2.3.0
SQLAlchemy==2.0.20
SQLAlchemy-Utils==0.41.1
psycopg2-binary==2.9.7
python-decouple==3.8
alembic==1.12.0
bcrypt==4.0.1
pyjwt==2.8.0